requirements.txt
Django>=5.2,<5.3
gunicorn>=23.0.0,<24.0.0
psycopg[binary]>=3.2.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
Pillow>=11.0.0,<12.0.0
whitenoise>=6.8.0,<7.0.0
